fix: deploy

+++ app/crud/base.py
@@ -57,8 +57,10 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         db.refresh(db_obj)
         return db_obj
 
-    def remove(self, db: Session, *, id: int) -> ModelType:
+    def remove(self, db: Session, *, id: int) -> Optional[ModelType]:
         obj = db.get(self.model, id)
+        if obj is None:
+            return None
         db.delete(obj)
         db.commit()
         return obj
